--- src/queryParameters.js
@@ -1,25 +1,0 @@
-import { useLocation, useHistory } from "react-router-dom";
-
-export const useQueryParameter = (key) => {
-  const location = useLocation();
-  const searchParams = new URLSearchParams(location.search);
-  return searchParams.get(key);
-};
-
-export const useReplaceQueryParameter = () => {
-  const location = useLocation();
-  const history = useHistory();
-  const searchParams = new URLSearchParams(location.search);
- 
-  return ({ basicURL, key, value }) => {
-
-    if (value === undefined) {
-      searchParams.delete(key);
-    } else {
-      searchParams.set(key, value);
-    }
-    
-    const newSearch = searchParams.toString();
-    history.push(`${basicURL}?${newSearch}`);
-  };
-};
